extract step tests
